backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Vector Database
qdrant-client==1.12.1

# ML - Core
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
pillow>=10.2.0
numpy>=1.26.0,<2.0.0

# ML - Face Recognition (optional - uses MTCNN + InceptionResnetV1)
facenet-pytorch
scikit-learn>=1.5.0

# Utilities
aiofiles==24.1.0
python-magic-bin==0.4.14
httpx==0.28.1
sse-starlette==2.1.0
